PYTHON/testApp/requirements.txt
# asgiref==3.7.2
django==4.2.6
djangorestframework==3.14.0
Pillow==10.1.0
PyJWT==1.7.1
pytz==2023.3.post1
# sqlparse==0.4.4
# tzdata==2023.3
